dts: sif: disable rtc2 alarm wakeup property
maxim,enable-rtc2-alarm-wakeup is causing system to
reboot when the alarm expires. In case of SW initiated
shutdown, this alarm is disabled. But in case of HW initiated
shutdown like soctherm and long on key press, this alarm remains
on and cause system reboot if power jack is kept connected.

bug 200518244
bug 2596581

Change-Id: Idd6bf205783c34b76153e000e6307a0d4b656808
Reviewed-on: https://git-master.nvidia.com/r/2119192
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>